make reorganizer script "auto suficient"
